Sucky Function call parse code. Not quite complete..
Went onto function call parsing. Not happy with this... I think I was
partially closer in the first attempt
as it actually allowed throwing errors AT the token that was an issue.

Thinking of stripping back the parser a little again and work partially
as I was but with what I've learnt so far about this. I need to handle
an identifier then figure out what I'm parsing... I need
to figure out where the error is though. can't do that currently.